Deploy fix

+++ devops/deploy.sh
@@ -16,9 +16,13 @@ nvm use default
 
 pm2 stop media-tray-update-server
 
+$DIRECTORY="media-tray-update-server"
 # keep previous copy in case we have to fall back
-rm -rf media-tray-update-server-prev
-mv media-tray-update-server media-tray-update-server-prev
+if [ -d "$DIRECTORY" ]; then
+  echo "$DIRECTORY does exist."
+  rm -rf media-tray-update-server-prev
+  mv media-tray-update-server media-tray-update-server-prev
+fi
 
 # assumes prev deploy task has unziped into folder called "media-tray-update-server-prep"
 mv media-tray-update-server-prep media-tray-update-server
